refactor: controller 에서 도메인 객체를 response 로 변경하도록 수정
